Initial commit
Bootstrap a re-frame app

Add a vexflow div that's totally separate from the reagent
stuff (vexflow is similar to d3 in that it wants direct dom access)
and hook up a very simple demo.

Hook in a very simple leipzig / cljs-bach demo. At the moment the
leipzig demo isn't using melody, it's using hand-generated maps,
but it's instructive and works